updated CloudFormation not to remove Cloud9

+++ deleteCloudFormationStacks.sh
@@ -1,5 +1,11 @@
 for stack in `aws cloudformation list-stacks --output text | cut -f5`
 do 
-   echo "Removing CloudFormation stack $stack"
-   aws cloudformation delete-stack --stack-name "${stack}"
+   echo "Processing CloudFormation stack $stack"
+  if [[ "$stack" != *"cloud9"* ]];
+	then
+		echo " Removing $stack"
+		aws cloudformation delete-stack --stack-name "${stack}"
+   else
+	echo "Not removing cloud9 $stack"
+  fi
 done
